+++ providers/src/airflow/providers/slack/hooks/slack.py
@@ -24,12 +24,11 @@ from functools import cached_property
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, TypedDict
 
-from deprecated import deprecated
 from slack_sdk import WebClient
 from slack_sdk.errors import SlackApiError
 from typing_extensions import NotRequired
 
-from airflow.exceptions import AirflowNotFoundException, AirflowProviderDeprecationWarning
+from airflow.exceptions import AirflowNotFoundException
 from airflow.hooks.base import BaseHook
 from airflow.providers.slack.utils import ConnectionExtraConfig
 from airflow.utils.helpers import exactly_one
@@ -185,67 +184,6 @@ class SlackHook(BaseHook):
             iterated on to execute subsequent requests.
         """
         return self.client.api_call(api_method, **kwargs)
-
-    @deprecated(
-        reason=(
-            "This method utilise `files.upload` Slack API method which is being sunset on March 11, 2025. "
-            "Beginning May 8, 2024, newly-created apps will be unable to 'files.upload' Slack API. "
-            "Please use `send_file_v2` or `send_file_v1_to_v2` instead."
-        ),
-        category=AirflowProviderDeprecationWarning,
-    )
-    def send_file(
-        self,
-        *,
-        channels: str | Sequence[str] | None = None,
-        file: str | Path | None = None,
-        content: str | None = None,
-        filename: str | None = None,
-        filetype: str | None = None,
-        initial_comment: str | None = None,
-        title: str | None = None,
-        **kwargs,
-    ) -> SlackResponse:
-        """
-        Create or upload an existing file.
-
-        :param channels: Comma-separated list of channel names or IDs where the file will be shared.
-            If omitting this parameter, then file will send to workspace.
-        :param file: Path to file which need to be sent.
-        :param content: File contents. If omitting this parameter, you must provide a file.
-        :param filename: Displayed filename.
-        :param filetype: A file type identifier.
-        :param initial_comment: The message text introducing the file in specified ``channels``.
-        :param title: Title of file.
-
-        .. seealso::
-            - `Slack API files.upload method <https://api.slack.com/methods/files.upload>`_
-            - `File types <https://api.slack.com/types/file#file_types>`_
-        """
-        if not exactly_one(file, content):
-            raise ValueError("Either `file` or `content` must be provided, not both.")
-        elif file:
-            file = Path(file)
-            with open(file, "rb") as fp:
-                if not filename:
-                    filename = file.name
-                return self.client.files_upload(
-                    file=fp,
-                    filename=filename,
-                    filetype=filetype,
-                    initial_comment=initial_comment,
-                    title=title,
-                    channels=channels,
-                )
-
-        return self.client.files_upload(
-            content=content,
-            filename=filename,
-            filetype=filetype,
-            initial_comment=initial_comment,
-            title=title,
-            channels=channels,
-        )
 
     def send_file_v2(
         self,